Move version tests from overlay to file actions.
